requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]>=2.7.0
structlog==24.1.0
pytest==8.2.0
pytest-cov==5.0.0
hypothesis==6.100.0
ultralytics==8.2.0
shapely==2.0.4
opencv-python-headless==4.9.0.80
streamlit==1.35.0
httpx==0.27.0
redis[asyncio]==5.0.4
python-jose[cryptography]==3.3.0
aioboto3==12.3.0
reportlab==4.2.2
openpyxl==3.1.2
apscheduler==3.10.4
